requirements.txt
Flask
celery
redis
pyyaml
bleach
dkimpy
pydmarc
cryptography
